test: factor run with ctl channel out of helper
