+++ test/helper.js
@@ -6,7 +6,7 @@
 'use strict';
 
 // test globals
-/* global assert,debug,fs,path,util,exec */
+/* global assert,util,exec */
 /* eslint-disable */
 global.assert = require('assert');
 global.debug = require('./debug');
@@ -18,8 +18,6 @@ global.util = require('util');
 require('shelljs/global');
 
 // module locals
-var child = require('child_process');
-var control = require('strong-control-channel/process');
 var supervise = require('./supervise');
 
 exports.runCtl = {
@@ -81,41 +79,3 @@ function pause(secs) {
 }
 
 global.pause = pause;
-
-exports.runWithControlChannel = function(appWithArgs, runArgs, onMessage) {
-  if (onMessage === undefined && typeof runArgs === 'function') {
-    onMessage = runArgs;
-    runArgs = [];
-  }
-
-  var ctl = path.resolve(path.dirname(appWithArgs[0]), 'runctl');
-  try {
-    fs.unlinkSync(ctl);
-  } catch (er) {
-    console.log('# no `%s` to cleanup: %s', ctl, er);
-  }
-
-  var options = {
-    stdio: [0, 1, 2, 'ipc'],
-    env: util._extend({
-      STRONGLOOP_BASE_INTERVAL: 500,
-      STRONGLOOP_FLUSH_INTERVAL: 2,
-    }, process.env),
-  };
-
-  var runner = require.resolve('../bin/sl-run');
-
-  var args = [
-    runner,
-    '--no-timestamp-workers',
-    '--no-timestamp-supervisor'
-  ].concat(runArgs).concat(appWithArgs);
-
-  debug('spawn: args=%j', args);
-
-  var c = child.spawn(process.execPath, args, options);
-  c.control = control.attach(onMessage, c);
-  c.unref();
-  c._channel.unref(); // There is no documented way to unref child IPC
-  return c;
-};
