+++ test/test-run-trace.js
@@ -8,7 +8,7 @@
 var debug = require('./debug');
 var helper = require('./helper');
 var http = require('http');
-var run = helper.runWithControlChannel;
+var run = require('./run-with-ctl-channel');
 var tap = require('tap');
 
 var skipIfNoLicense = process.env.STRONGLOOP_LICENSE
